Improve flow calculation of model towards better supply/demand modelling
Previously, the model would get stuck if there is e.g. high demand for
recycled materials, but no supply of it. The difference between supply
and demand was not covered e.g. by extracting additional naturla
resources. This update to the model works around the issue as follows:

1. Flows try to fulfill the combined demand of a good, not just the
partial demand that is derived from the respective incentive. This
allows overproduction in one area to cover underproduction in another.
Since overproductions are used up eventually, the supply and demand will
match up at some point (if there are enough resources).
2. The flow from natural resources can theoretically be unbounded.
However, if the incentive for using natural resources is small compared
to the recycling incentive, natural resources would not be used, even if
the supply of recycled materials is zero. Therefore, the flow from
natural resources will now always fullfil the demand that is not covered
by recycled materials. This behavior is translated to the newly produced
phones: This flow will try to cover the demand that is not covered by
the reuse, repair and refurbish flows, such that new phones can always
enter the system and become phones in use.