Replace happiness score with coverage score
The happiness (formerly user satisfactiin) maxed out at 100% and would
shrink again if the would be more phones in used than needed. Since the
wording of this score changed to "Demand met", the score is renamed to
`coverage` and is now just be the ratio of `phonesInUse` and
`phoneGoal`, i.e. it is possible to have > 100% coverage.