Add capacity and supply mechanism for natural resources mining
Recent changes to the model changed the flows to compensate
undersupplies in some areas with oversupplies in others. This causes
trouble for the natural resources, which are an unbounded resource, at
least from the simplified world view of the model. This patch add a
new stock for natural resource supply and combines it with the usual
demand and capacity mechanisms. An additional parameter
`naturalResourceMiningRate` is introduced such that the speed of
resources mining can be controlled independently of the production rate.