Work around issue where parameter transforms are lost
The model simulator steps the model forward in time and has its own
internal state. However, the model store keep track of the current
record and it serves as a reference to the whole app. The simulator
needs to be in sync with the global state. This fix ensures that the
currently active set of parameters is preserved during sync.