Fix control panel not being hidable after switching to flex layout
A previous patch switched the `display` property of the control panel to
`flex` layout. However, the visibility of the control panel is
controlled by a CSS class that also sets the `display` property, but
this style change is overridden by settjng the `flex` layout in a style
file with higher priority. This patch wraps the control panel into
another div that can be utilized for the visibility switching.