Fix bug that prevented certain layer conditions to be reevaluated
Layer conditions that do not contain references to the current record
did not trigger the reactive reevaluation of the condition when the
record is updated. The condition was only evaluated once upon app
restart. While this saves some computation time for constant conditions,
it breaks conditions that depend on non-constant global state like the
current date or time.

With this patch, conditions always access the record timestap
internally during evaluation, which should fix the issue for all
situations where the simulation progresses in time such as updates via
the animation loop.