Add `reset` property to conditional layer
When an animated conditional layers are hidden, it continues the
animation in the background. When it is shown again, the animation would
not start from the beginning, but rather from frame it progressed to
while it was hidden.

This patch introduces a `reset` propery for conditional layers, that,
when set to `true`, causes the layer to unload whenever it is hidden.
When the layer is added back again, the animation we start from the
beginning. The default value of `reset` is `false`.