Fix animation resetting in Chrome on Linux
Using `v-if` to unmount and destroy the conditional layer component
doesn't reliably reset animations in Chrome on Linux. The new approach
does not use `v-if`, but rather sets the `src` attribute of the image to
a base64 encoded single-pixel transparent PNG when the layer is
deactivaged. The issue could not be reproduced usinfg this new method.